[#55] refactor: event 인터페이스 및 구현체 구현
인터페이스와 메시지 타입만 application layer 에 작성. 구현체는 infrastructure 에 작성한다.

consumer 는 인터페이스 조차도 없음. 왜냐면 consumer 는 외부 서비스에서 처리한다는 가정하에 만들었기 때문. (마이크로서비스)
