+++ src/main/java/com/blog/bespoke/application/event/message/UserRegistrationMessage.java
@@ -1,16 +1,12 @@
-package com.blog.bespoke.application.amqp.message;
+package com.blog.bespoke.application.event.message;
 
 import lombok.Builder;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 import java.io.Serializable;
 
-@Getter
-@Setter
-@Builder
 @Data
+@Builder
 public class UserRegistrationMessage implements Serializable {
     private String email;
     private String code;
