+++ src/main/java/com/blog/bespoke/application/event/publisher/EventPublisher.java
@@ -1,0 +1,7 @@
+package com.blog.bespoke.application.event.publisher;
+
+import com.blog.bespoke.application.event.message.UserRegistrationMessage;
+
+public interface EventPublisher {
+    void publishMailEvent(UserRegistrationMessage message);
+}
