--- src/main/java/com/blog/bespoke/application/event/publisher/UserRegistrationEventPublisher.java
@@ -1,19 +1,0 @@
-package com.blog.bespoke.application.event.publisher;
-
-import com.blog.bespoke.application.event.type.UserRegistrationEvent;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class UserRegistrationEventPublisher {
-    private final ApplicationEventPublisher publisher;
-
-    public void publish(String email, String code) {
-        publisher.publishEvent(UserRegistrationEvent.builder()
-                .email(email)
-                .code(code)
-                .build());
-    }
-}
