+++ src/main/java/com/blog/bespoke/infrastructure/event/rabbitmq/publisher/RabbitMqPublisher.java
@@ -1,14 +1,13 @@
-package com.blog.bespoke.application.amqp.publisher;
+package com.blog.bespoke.infrastructure.event.rabbitmq.publisher;
 
-import com.blog.bespoke.application.amqp.message.UserRegistrationMessage;
+import com.blog.bespoke.application.event.message.UserRegistrationMessage;
+import com.blog.bespoke.application.event.publisher.EventPublisher;
 import lombok.RequiredArgsConstructor;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
 
-@Service
 @RequiredArgsConstructor
-public class EventPublisher {
+public class RabbitMqPublisher implements EventPublisher {
     private final RabbitTemplate rabbitTemplate;
 
     @Value("${rabbitmq.routing-key.mail}")
